Change upload to format, add nav link if import screen viewed
