--- src/lib/components/event/event-json.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-  import CodeBlock from '$lib/components/code-block.svelte';
-
-  export let events: HistoryEventWithId[];
-</script>
-
-<CodeBlock content={events} />
