--- src/lib/stores/uploads-events.ts
@@ -1,6 +1,0 @@
-import { writable } from 'svelte/store';
-
-export const uploadEvents = writable<FetchEventsResponse>({
-  events: [],
-  eventGroups: [],
-});
